+++ SheeeeeeeeetExample/Appearance/DemoColors.swift
@@ -1,0 +1,18 @@
+//
+//  DemoColors.swift
+//  SheeeeeeeeetExample
+//
+//  Created by Daniel Saidi on 2019-08-09.
+//  Copyright © 2019 Daniel Saidi. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    
+    static var demoBlue: UIColor { return UIColor(hex: 0x0FA2F5) }
+    static var demoGreen: UIColor { return UIColor(hex: 0x81c03f) }
+    static var demoPink: UIColor { return UIColor(hex: 0xec5f72) }
+    static var demoPurple: UIColor { return UIColor(hex: 0xd9007b) }
+    static var demoRed: UIColor { return UIColor(hex: 0xff3333) }
+}